Move Button, Text, and TextField demos to Components demo
