+++ app/src/commonMain/kotlin/com/alexrdclement/uiplayground/app/demo/components/demo/TextStyle.kt
@@ -1,0 +1,15 @@
+package com.alexrdclement.uiplayground.app.demo.components.demo
+
+enum class TextStyle {
+    Headline,
+    Display,
+    TitleSmall,
+    TitleMedium,
+    TitleLarge,
+    BodyLarge,
+    BodyMedium,
+    BodySmall,
+    LabelLarge,
+    LabelMedium,
+    LabelSmall,
+}
