+++ app/src/commonMain/kotlin/com/alexrdclement/uiplayground/app/demo/experiments/Experiment.kt
@@ -5,9 +5,6 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 enum class Experiment : CatalogItem {
-    Button,
-    Text,
-    TextField,
     Gradients;
 
     override val title = this.name
