+++ app/src/commonMain/kotlin/com/alexrdclement/uiplayground/app/demo/experiments/ExperimentScreen.kt
@@ -4,10 +4,7 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import com.alexrdclement.uiplayground.app.configuration.ConfigureButton
-import com.alexrdclement.uiplayground.app.demo.experiments.demo.button.ButtonDemo
 import com.alexrdclement.uiplayground.app.demo.experiments.demo.gradient.GradientDemo
-import com.alexrdclement.uiplayground.app.demo.experiments.demo.text.TextDemo
-import com.alexrdclement.uiplayground.app.demo.experiments.demo.textfield.TextFieldDemo
 import com.alexrdclement.uiplayground.components.BackNavigationButton
 import com.alexrdclement.uiplayground.components.Scaffold
 import com.alexrdclement.uiplayground.components.Text
@@ -32,9 +29,6 @@ fun ExperimentScreen(
         },
     ) { innerPadding ->
         when (experiment) {
-            Experiment.Button -> ButtonDemo()
-            Experiment.Text -> TextDemo()
-            Experiment.TextField -> TextFieldDemo()
             Experiment.Gradients -> GradientDemo(modifier = Modifier.padding(innerPadding))
         }
     }
